06-Jan-2026/FastAPI-CRUD/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
dnspython==1.16.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==3.12.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
python-dotenv
pytest
httpx
mongomock